flum: General modeling function for Attributable Influenza Mortality

General function wrapper to perform a variety of forecasted model. The output will always be a set of fitted values.

flum(data=NULL, model="serflm", outc=NULL, epi=NULL, time=NULL,
              t.interval=52, echo=F)

Arguments

data
A dataframe class object, must contain time variable, epidemic indicator, and measure of influenza morbidity

outc
an unquoted name of a column in data which corresponds to the outcome variable of interest

epi
an unquoted name of a column in data object (e.g. epi) or if null will default to Sept - May.

time
an unquoted name of a column in data object, must be a numeric/integer class variable in dataframe which corresponds to a unit of time, must be unique (i.e. non-repeating)

t.interval
a numeric vector indicating period length, i.e. 52 weeks in year

echo
A logical parameter, if T Will output model parameters and description.

...
Individual models may need additional parameters, e.g. for virology must specify a viral test variable.

model
A character of length 1, indicates which model to fit. Options include "ird" (see ?ird), "serflm" (see ?serflm), virology model (vignette('05-modserf', package = "flumodelr")), "arima" (see ?flurima).

Value

an object of class data.frame, input + y0 (fitted values), threshold if applicable, and computted excess mortality.

References

Thompson WW1, Weintraub E, Dhankhar P, Cheng PY, Brammer L, Meltzer MI, Bresee JS, Shay DK. Estimates of US influenza-associated deaths made using four different methods. Influenza Other Respir Viruses. 2009 Jan;3(1):37-49. /urlhttps://www.ncbi.nlm.nih.gov/pubmed/19453440

Examples

require(flumodelr)
flu_ex <- flumodelr::flu_ex
flu_fit <- flum(flu_ex, model="serflm", outc = fludeaths, time = yrweek_dt)

head(flu_fit)

#> NULL