serflm: Fit a Serling Model on Time Series Data

Performs a cyclical linear regression model, or 'serfling' model

serflm(data = NULL, outc = NULL, epi = NULL, time = NULL,
  t.interval = "wofy")

Arguments

data
A dataframe class object, must contain time variable, epidemic indicator, and measure of influenza morbidity

outc
A character string of length=1, identifies a numeric type variable in dataframe data which is the measure of disease morbidity / mortality

epi
either a character string of length=1 (e.g. '"epi")', identifying a logical type variable in dataframe data which indicates epidemic vs non-epidemic periods, or a vector of length=2 which identifies periods of epidemicity, e.g. 'c(40,20)'.

time
A character string of length=1, identifies a numeric/integer class variable in dataframe which corresponds to a unit of time, must be unique (i.e. non-repeating)

t.interval
A character string, must specify whether unit of time cycle is weeks/years ("wofy", i.e. 52 weeks), or month/years ("mofy"), days/years("dofy")

Value

an object of class data.frame (input dataframe + fitted values)

References

Serfling RE. Methods for current statistical analysis of excess pneumonia-influenza deaths. Public Health Rep. 1963 Jun; 78(6): 494 - 506. /urlhttps://www.ncbi.nlm.nih.gov/pmc/articles/PMC1915276/

Examples

require(flumodelr)
flu_ex <- flumodelr::flu_ex
fit <- serflm(data=flu_ex, outc = "fludeaths", epi="epi", time="yrweek_dt")

#> Cyclical Regression Model 
#>  ============================================================ 
#> Setting regression parameters...
#>  'epi' variable is: epi 
#>  'time' variable is: yrweek_dt 
#>   time interval is: 52 
#> Model Formula: 
#> fludeaths ~ t_unit + sin_f1 + cos_f1
#> t_unit is:  yrweek_dt 
#>  sin_f1 / cos_f1 is a Fourier term

#> Error in lm.fit(x, y, offset = offset, singular.ok = singular.ok, ...): 0 (non-NA) cases

summary(fit)

#>       year           week         fludeaths        yrweek_dt         
#>  Min.   :2010   Min.   : 1.00   Min.   : 4.083   Min.   :2010-10-18  
#>  1st Qu.:2012   1st Qu.:14.00   1st Qu.: 5.368   1st Qu.:2012-01-12  
#>  Median :2013   Median :27.00   Median : 6.203   Median :2013-04-22  
#>  Mean   :2013   Mean   :26.59   Mean   : 6.524   Mean   :2013-04-15  
#>  3rd Qu.:2014   3rd Qu.:39.50   3rd Qu.: 7.296   3rd Qu.:2014-07-17  
#>  Max.   :2015   Max.   :52.00   Max.   :11.808   Max.   :2015-10-12  
#>   prop_flupos           week2           theta             sin_f1       
#>  Min.   :0.001087   Min.   :  1.0   Min.   :0.03846   Min.   :-1.0000  
#>  1st Qu.:0.021758   1st Qu.: 65.5   1st Qu.:2.51923   1st Qu.:-0.7058  
#>  Median :0.057316   Median :130.0   Median :5.00000   Median : 0.0000  
#>  Mean   :0.099052   Mean   :130.0   Mean   :5.00000   Mean   : 0.0000  
#>  3rd Qu.:0.154002   3rd Qu.:194.5   3rd Qu.:7.48077   3rd Qu.: 0.7058  
#>  Max.   :0.383560   Max.   :259.0   Max.   :9.96154   Max.   : 1.0000  
#>      cos_f1              pred_y     
#>  Min.   :-1.000000   Min.   :4.557  
#>  1st Qu.:-0.705817   1st Qu.:5.402  
#>  Median : 0.000000   Median :6.578  
#>  Mean   :-0.003861   Mean   :6.524  
#>  3rd Qu.: 0.663123   3rd Qu.:7.640  
#>  Max.   : 1.000000   Max.   :8.453