fluglm: Generate fitted prediction values with generalized linear model.

Similar to serflm, but allows for more specifications.

fluglm(data=NULL, outc=NULL, epi=NULL, time=NULL,
              period=52, echo=F, poly=T, model.form=NULL,
              int_type="ci", alpha=0.1)

Arguments

data
A dataframe class object, must contain time variable, epidemic indicator, and measure of influenza morbidity

outc
an unquoted name of a column in data which corresponds to the outcome variable of interest

bl_type
A quote string either: "season" or "viral". Denoting type of baseline to use for fitting model.

bl_var
an unquoted name of a variable corresponding to bl_type.

time
an unquoted name of a column in data object, must be a numeric/integer class variable in dataframe which corresponds to a unit of time, must be unique (i.e. non-repeating)

period
a numeric vector indicating period length, i.e. 52 weeks in year

echo
A logical parameter, if T will print variables used in model.

poly
A logical parameter, if T will include a quadratic, cubic and quartic term.

model_form
An object of type formula, allowing for user-specified model to be passed on to glm(). Default missing.

int_type
Specifies type of upper interval to be output, currently only allows for confidence intervals. Predictional intervals to be added, but are only approximate for Poisson families.

alpha
The threshold for interval, default is 0.05 (one-sided).

offset
Specify if offset term to be used, must specify log(object)

...
options passed on to glm model (e.g. family=poisson, see ?glm)

Value

an object of class data.frame, input + y0 (fitted values), y0_ul the upper threshold

References

Thompson WW1, Weintraub E, Dhankhar P, Cheng PY, Brammer L, Meltzer MI, Bresee JS, Shay DK. Estimates of US influenza-associated deaths made using four different methods. Influenza Other Respir Viruses. 2009 Jan;3(1):37-49. /urlhttps://www.ncbi.nlm.nih.gov/pubmed/19453440

Examples

require(flumodelr)
fludta <- flumodelr::fludta
flu_fit <- fluglm(fludta, outc = fludeaths, time = yrweek_dt)

head(flu_fit)

#> # A tibble: 6 x 10
#>    year  week fludeaths alldeaths perc_fl~ yrweek_dt  prop_f~ epi      y0 y0_ul
#>   <int> <int>     <int>     <int>    <dbl> <date>       <dbl> <lgl> <dbl> <dbl>
#> 1  2010    42       726     11269     6.44 2010-10-17  0.0133 T       805   946
#> 2  2010    43       720     11204     6.43 2010-10-24  0.0193 T       815   951
#> 3  2010    44       739     11115     6.65 2010-10-31  0.0173 T       826   958
#> 4  2010    45       724     10321     7.01 2010-11-07  0.0231 T       838   966
#> # ... with 2 more rows