ird: Perform incidence rate-difference models

Calculates the rate difference between influenza season and peri-influenza season, or between influenza season and summer season

ird(data=NULL, outc=NULL, time=NULL, viral,
            period=52, respStart=27, high=0.1, fluStart=40,
            fluStop=18, echo=F)

Arguments

data
A dataframe class object, must contain time variable, epidemic indicator, and measure of influenza morbidity

outc
A named variable in 'data' dataframe, identifies a numeric type variable in dataframe data which is the measure of disease morbidity / mortality

time
A named variable in 'data' dataframe, identifies a numeric/integer class variable in dataframe which corresponds to a unit of time, must be unique (i.e. non-repeating)

viral
A named variable in 'data' data.frame, identifies a numeric type variable in dataframe data which is the measure of viral activity.

period
a numeric variable indicating the period length of the time variable (i.e. weeks = 52, the default)

respStart
A numeric/integer class variable, must specify the week number of the start of the respiratory season. The default value is 27. This corresponds with the beginning of July.

high
A numeric class variable, must specify the proportion of positive isolates for a week to be considered of "high viral activity". The default value is 0.10.

fluStart
A numeric/integer class variable, must specify the week number of the start of the influenza season. The default value is 40. This corresponds with the beginning of October.

fluStop
A numeric/integer class variable, must specify the week number of the start of the influenza season. The default value is 18. This corresponds with the end of April.

Value

an object of class data.frame

References

Thompson WW, Weintraub E, Dhankhar P, Cheng PY, Brammer L, Meltzer MI, Bresee JS, Shay DK. Estimates of US influenza‐associated deaths made using four different methods. Influenza and other respiratory viruses. 2009 Jan 1;3(1):37-49. /urlhttp://onlinelibrary.wiley.com/doi/10.1111/j.1750-2659.2009.00073.x/full

Examples

require(flumodelr)
fludta <- flumodelr::fludta
flu_fit <- ird(data=fludta, outc = perc_fludeaths, viral=prop_flupos, time=yrweek_dt)
flu_rates <- rb(flu_fit, perc_fludeaths)

#> Warning: Column `fluseason`/`high` has different attributes on LHS and RHS of join

flu_rates

#> # A tibble: 12 x 4
#> # Groups:   season [?]
#>    season high_act perc_fludeaths_fluseason perc_fludeaths_viral_act
#>     <dbl> <lgl>                       <dbl>                    <dbl>
#>  1   2010 FALSE                        7.00                     6.83
#>  2   2010 TRUE                         7.54                     8.05
#>  3   2011 FALSE                        6.19                     6.47
#>  4   2011 TRUE                         6.96                     7.06
#>  5   2012 FALSE                        6.01                     6.10
#>  6   2012 TRUE                         7.27                     7.67
#>  7   2013 FALSE                        5.76                     5.75
#>  8   2013 TRUE                         6.81                     7.19
#>  9   2014 FALSE                        5.74                     5.71
#> 10   2014 TRUE                         6.79                     7.37
#> 11   2015 FALSE                        5.50                     5.51
#> 12   2015 TRUE                         5.54                    NA