serflm: Fit a Serling Model on Time Series Data

Performs a cyclical linear regression model, or 'serfling' model

serflm(data = NULL, outc = NULL, epi = NULL, time = NULL,
  t.interval = 52, echo = F)

Arguments

data
A dataframe class object, must contain time variable, epidemic indicator, and measure of influenza morbidity

outc
an unquoted name of a column in data which corresponds to the outcome variable of interest

epi
an unquoted name of a column in data object (e.g. epi) or if null will default to Sept - May.

time
an unquoted name of a column in data object, must be a numeric/integer class variable in dataframe which corresponds to a unit of time, must be unique (i.e. non-repeating)

t.interval
a numeric vector indicating period length, i.e. 52 weeks in year

echo
A logical parameter, if T. Will print variables used in model.

Value

an object of class data.frame, input + y0 (fitted values), y0_ul the upper serfling threshold

References

Serfling RE. Methods for current statistical analysis of excess pneumonia-influenza deaths. Public Health Rep. 1963 Jun; 78(6): 494 - 506. /urlhttps://www.ncbi.nlm.nih.gov/pmc/articles/PMC1915276/

Examples

require(flumodelr)
flu_ex <- flumodelr::flu_ex
flu_fit <- serflm(flu_ex, outc = fludeaths, time = yrweek_dt)

head(flu_fit)

#> # A tibble: 6 x 8
#>    year  week fludeaths  yrweek_dt prop_flupos   epi       y0    y0_ul
#>   <int> <int>     <dbl>     <date>       <dbl> <lgl>    <dbl>    <dbl>
#> 1  2010    42  6.482143 2010-10-18  0.01331601  TRUE 6.325176 7.419003
#> 2  2010    43  6.605505 2010-10-25  0.01928211  TRUE 6.415187 7.509014
#> 3  2010    44  6.779817 2010-11-01  0.01726122  TRUE 6.507871 7.601698
#> 4  2010    45  6.581818 2010-11-08  0.02305556  TRUE 6.601823 7.695650
#> # ... with 2 more rows